Move metrics publishers to their own directory
