+++ llarp/metrics/publishers.cpp
@@ -1,4 +1,4 @@
-#include <util/metrics_publishers.hpp>
+#include <metrics/publishers.hpp>
 
 #include <fstream>
 #include <iostream>
