+++ test/metrics/test_llarp_metrics_publisher.cpp
@@ -1,4 +1,4 @@
-#include <util/metrics_publishers.hpp>
+#include <metrics/publishers.hpp>
 
 #include <gtest/gtest.h>
 #include <gmock/gmock.h>
